Add LedgerValidation to perform LedgerOperations calculation
