+++ test/archethic/mining/validation_context_test.exs
@@ -7,6 +7,7 @@ defmodule Archethic.Mining.ValidationContextTest do
   alias Archethic.Election
 
   alias Archethic.Mining.Fee
+  alias Archethic.Mining.LedgerValidation
   alias Archethic.Mining.ValidationContext
 
   alias Archethic.P2P
@@ -543,12 +544,13 @@ defmodule Archethic.Mining.ValidationContextTest do
     resolved_addresses = Enum.map(movements, &{&1.to, &1.to}) |> Map.new()
 
     ledger_operations =
-      %LedgerOperations{fee: fee}
-      |> LedgerOperations.filter_usable_inputs(unspent_outputs, contract_context)
-      |> LedgerOperations.mint_token_utxos(tx, timestamp, current_protocol_version())
-      |> LedgerOperations.build_resolved_movements(movements, resolved_addresses, tx.type)
-      |> LedgerOperations.validate_sufficient_funds()
-      |> LedgerOperations.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      %LedgerValidation{fee: fee}
+      |> LedgerValidation.filter_usable_inputs(unspent_outputs, contract_context)
+      |> LedgerValidation.mint_token_utxos(tx, timestamp, current_protocol_version())
+      |> LedgerValidation.build_resolved_movements(movements, resolved_addresses, tx.type)
+      |> LedgerValidation.validate_sufficient_funds()
+      |> LedgerValidation.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      |> LedgerValidation.to_ledger_operations()
 
     %ValidationStamp{
       timestamp: timestamp,
@@ -574,12 +576,13 @@ defmodule Archethic.Mining.ValidationContextTest do
     resolved_addresses = Enum.map(movements, &{&1.to, &1.to}) |> Map.new()
 
     ledger_operations =
-      %LedgerOperations{fee: fee}
-      |> LedgerOperations.filter_usable_inputs(unspent_outputs, contract_context)
-      |> LedgerOperations.mint_token_utxos(tx, timestamp, current_protocol_version())
-      |> LedgerOperations.build_resolved_movements(movements, resolved_addresses, tx.type)
-      |> LedgerOperations.validate_sufficient_funds()
-      |> LedgerOperations.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      %LedgerValidation{fee: fee}
+      |> LedgerValidation.filter_usable_inputs(unspent_outputs, contract_context)
+      |> LedgerValidation.mint_token_utxos(tx, timestamp, current_protocol_version())
+      |> LedgerValidation.build_resolved_movements(movements, resolved_addresses, tx.type)
+      |> LedgerValidation.validate_sufficient_funds()
+      |> LedgerValidation.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      |> LedgerValidation.to_ledger_operations()
 
     %ValidationStamp{
       timestamp: timestamp,
@@ -605,12 +608,13 @@ defmodule Archethic.Mining.ValidationContextTest do
     resolved_addresses = Enum.map(movements, &{&1.to, &1.to}) |> Map.new()
 
     ledger_operations =
-      %LedgerOperations{fee: fee}
-      |> LedgerOperations.filter_usable_inputs(unspent_outputs, contract_context)
-      |> LedgerOperations.mint_token_utxos(tx, timestamp, current_protocol_version())
-      |> LedgerOperations.build_resolved_movements(movements, resolved_addresses, tx.type)
-      |> LedgerOperations.validate_sufficient_funds()
-      |> LedgerOperations.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      %LedgerValidation{fee: fee}
+      |> LedgerValidation.filter_usable_inputs(unspent_outputs, contract_context)
+      |> LedgerValidation.mint_token_utxos(tx, timestamp, current_protocol_version())
+      |> LedgerValidation.build_resolved_movements(movements, resolved_addresses, tx.type)
+      |> LedgerValidation.validate_sufficient_funds()
+      |> LedgerValidation.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      |> LedgerValidation.to_ledger_operations()
 
     %ValidationStamp{
       timestamp: timestamp,
@@ -637,12 +641,13 @@ defmodule Archethic.Mining.ValidationContextTest do
     encoded_state = nil
 
     ledger_operations =
-      %LedgerOperations{fee: fee}
-      |> LedgerOperations.filter_usable_inputs(unspent_outputs, contract_context)
-      |> LedgerOperations.mint_token_utxos(tx, timestamp, current_protocol_version())
-      |> LedgerOperations.build_resolved_movements(movements, resolved_addresses, tx.type)
-      |> LedgerOperations.validate_sufficient_funds()
-      |> LedgerOperations.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      %LedgerValidation{fee: fee}
+      |> LedgerValidation.filter_usable_inputs(unspent_outputs, contract_context)
+      |> LedgerValidation.mint_token_utxos(tx, timestamp, current_protocol_version())
+      |> LedgerValidation.build_resolved_movements(movements, resolved_addresses, tx.type)
+      |> LedgerValidation.validate_sufficient_funds()
+      |> LedgerValidation.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      |> LedgerValidation.to_ledger_operations()
 
     %ValidationStamp{
       timestamp: timestamp,
@@ -729,12 +734,13 @@ defmodule Archethic.Mining.ValidationContextTest do
     encoded_state = nil
 
     ledger_operations =
-      %LedgerOperations{fee: fee}
-      |> LedgerOperations.filter_usable_inputs(unspent_outputs, contract_context)
-      |> LedgerOperations.mint_token_utxos(tx, timestamp, current_protocol_version())
-      |> LedgerOperations.build_resolved_movements(movements, resolved_addresses, tx.type)
-      |> LedgerOperations.validate_sufficient_funds()
-      |> LedgerOperations.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      %LedgerValidation{fee: fee}
+      |> LedgerValidation.filter_usable_inputs(unspent_outputs, contract_context)
+      |> LedgerValidation.mint_token_utxos(tx, timestamp, current_protocol_version())
+      |> LedgerValidation.build_resolved_movements(movements, resolved_addresses, tx.type)
+      |> LedgerValidation.validate_sufficient_funds()
+      |> LedgerValidation.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      |> LedgerValidation.to_ledger_operations()
 
     %ValidationStamp{
       timestamp: timestamp,
@@ -760,12 +766,13 @@ defmodule Archethic.Mining.ValidationContextTest do
     encoded_state = nil
 
     ledger_operations =
-      %LedgerOperations{fee: fee}
-      |> LedgerOperations.filter_usable_inputs(unspent_outputs, contract_context)
-      |> LedgerOperations.mint_token_utxos(tx, timestamp, current_protocol_version())
-      |> LedgerOperations.build_resolved_movements(movements, resolved_addresses, tx.type)
-      |> LedgerOperations.validate_sufficient_funds()
-      |> LedgerOperations.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      %LedgerValidation{fee: fee}
+      |> LedgerValidation.filter_usable_inputs(unspent_outputs, contract_context)
+      |> LedgerValidation.mint_token_utxos(tx, timestamp, current_protocol_version())
+      |> LedgerValidation.build_resolved_movements(movements, resolved_addresses, tx.type)
+      |> LedgerValidation.validate_sufficient_funds()
+      |> LedgerValidation.consume_inputs(tx.address, timestamp, encoded_state, contract_context)
+      |> LedgerValidation.to_ledger_operations()
       |> Map.put(
         :consumed_inputs,
         [
